토큰을 초기화할때 사용할 useToken을 만들고, 프로필에 적용 한글 프로필이 URL decode가 되지 않아서 정상적으로 찾지 못하던 점 수정
